WIP lazy loading older messages

+++ composeApp/src/commonMain/kotlin/ui/conversation/MessagesRemoteMediator.kt
@@ -6,12 +6,13 @@ import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import coil3.network.HttpException
 import data.io.base.BaseResponse
+import data.io.base.paging.MatrixPagingMetaIO
 import data.io.base.paging.PagingEntityType
-import data.io.base.paging.PagingMetaIO
+import data.io.matrix.room.event.content.constructMessages
 import data.io.social.network.conversation.message.ConversationMessageIO
 import data.io.social.network.conversation.message.ConversationMessagesResponse
 import database.dao.ConversationMessageDao
-import database.dao.PagingMetaDao
+import database.dao.matrix.MatrixPagingMetaDao
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
 import kotlinx.coroutines.withContext
@@ -24,18 +25,22 @@ import kotlinx.io.IOException
 @OptIn(ExperimentalPagingApi::class)
 class MessagesRemoteMediator (
     private val conversationMessageDao: ConversationMessageDao,
-    private val pagingMetaDao: PagingMetaDao,
+    private val pagingMetaDao: MatrixPagingMetaDao,
     private val conversationId: String,
     private val size: Int,
-    private val initialPage: Int = 0,
+    private val initialBatch: suspend () -> String?,
+    private val prevBatch: suspend () -> String?,
     private val invalidatePagingSource: () -> Unit,
-    private val getItems: suspend (page: Int) -> BaseResponse<ConversationMessagesResponse>,
+    private val getItems: suspend (batch: String?) -> BaseResponse<ConversationMessagesResponse>,
     private val cacheTimeoutMillis: Int = 24 * 60 * 60 * 1000
-): RemoteMediator<Int, ConversationMessageIO>() {
+): RemoteMediator<String, ConversationMessageIO>() {
+
+    val entityType: String
+        get() = "${PagingEntityType.ConversationMessage}_$conversationId"
 
     override suspend fun initialize(): InitializeAction {
         val timeElapsed = Clock.System.now().toEpochMilliseconds().minus(
-            pagingMetaDao.getCreationTime(PagingEntityType.ConversationMessage.name + conversationId) ?: 0
+            pagingMetaDao.getCreationTime(entityType) ?: 0
         )
 
         return if (timeElapsed < cacheTimeoutMillis) {
@@ -47,52 +52,54 @@ class MessagesRemoteMediator (
 
     override suspend fun load(
         loadType: LoadType,
-        state: PagingState<Int, ConversationMessageIO>
+        state: PagingState<String, ConversationMessageIO>
     ): MediatorResult {
-        var page: Int = when (loadType) {
-            LoadType.REFRESH -> {
-                getPagingMetaClosestToCurrentPosition(state)?.nextPage ?: initialPage
-            }
-            LoadType.PREPEND -> {
-                val pagingMeta = getPagingMetaForFirstItem(state)
-                val prevKey = if((pagingMeta?.currentPage ?: initialPage) > 0) pagingMeta?.currentPage?.minus(1) else null
-                prevKey ?: return MediatorResult.Success(endOfPaginationReached = pagingMeta != null)
-            }
-            LoadType.APPEND -> {
-                val pagingMeta = getPagingMetaForLastItem(state)
-                pagingMeta?.nextPage ?: return MediatorResult.Success(endOfPaginationReached = pagingMeta != null)
-            }
+        var batch: String? = when (loadType) {
+            LoadType.REFRESH -> getPagingMetaClosestToCurrentPosition(state)?.currentBatch
+            LoadType.PREPEND -> getPagingMetaForFirstItem(state)?.prevBatch ?: prevBatch()
+            LoadType.APPEND -> getPagingMetaForLastItem(state)?.nextBatch
         }
 
         try {
-            val apiResponse = getItems.invoke(if(loadType == LoadType.REFRESH) initialPage else page)
+            val data = getItems.invoke(batch).success?.data
 
-            val items = apiResponse.success?.data?.content
-            val endOfPaginationReached = items?.size != size
+            val items = constructMessages(
+                state = data?.state.orEmpty(),
+                timeline = data?.chunk.orEmpty(),
+                roomId = conversationId,
+                prevBatch = data?.end,
+                nextBatch = data?.start,
+                currentBatch = batch
+            )
+            println("kostka_test, saving ${items.size} messages under $batch")
+            val endOfPaginationReached = items.size != size
 
             return withContext(Dispatchers.IO) {
-                if (loadType == LoadType.REFRESH) {
-                    page = initialPage
-                    pagingMetaDao.removeAll()
-                    //conversationMessageDao.removeAll(conversationId)
-                }
-                val prevKey = if (page > 1) page - 1 else null
-                val nextKey = if (endOfPaginationReached) null else page + 1
-                items?.map {
-                    PagingMetaIO(
-                        entityId = conversationId,
-                        previousPage = prevKey,
-                        entityType = PagingEntityType.ConversationMessage.name + conversationId,
-                        currentPage = page,
-                        nextPage = nextKey
+                if (loadType == LoadType.REFRESH) pagingMetaDao.removeAll()
+
+                items.map {
+                    MatrixPagingMetaIO(
+                        entityId = it.id,
+                        prevBatch = data?.end,
+                        nextBatch = data?.start,
+                        currentBatch = batch,
+                        entityType = entityType
                     )
-                }?.let {
+                }.also {
                     pagingMetaDao.insertAll(it)
-                    conversationMessageDao.insertAll(items.onEach { item ->
-                        item.conversationId = conversationId
-                    })
-                    invalidatePagingSource()
                 }
+                // TODO the reason why data appeared and disappeared right after is because last batch will be null,
+                //  yet the first one as well, we should identify them in a different way
+                // first data, we finally get to know the current batch with a hack of sorts
+                println("kostka_test, batch: $batch, initial: ${initialBatch()}, updating to ${data?.start}")
+                if(batch == initialBatch()) {
+                    conversationMessageDao.setInitialBatch(
+                        conversationId = conversationId,
+                        batch = data?.start
+                    )
+                }
+                conversationMessageDao.insertAll(items)
+                invalidatePagingSource()
                 MediatorResult.Success(endOfPaginationReached = endOfPaginationReached)
             }
         } catch (exception: IOException) {
@@ -103,28 +110,28 @@ class MessagesRemoteMediator (
     }
 
     /** Returns paging meta data from the current position item */
-    private suspend fun getPagingMetaClosestToCurrentPosition(state: PagingState<Int, ConversationMessageIO>): PagingMetaIO? {
+    private suspend fun getPagingMetaClosestToCurrentPosition(state: PagingState<String, ConversationMessageIO>): MatrixPagingMetaIO? {
         return state.anchorPosition?.let { position ->
-            state.closestItemToPosition(position)?.conversationId?.let { id ->
+            state.closestItemToPosition(position)?.id?.let { id ->
                 pagingMetaDao.getByEntityId(id)
             }
         }
     }
 
     /** Returns paging meta data from the first item */
-    private suspend fun getPagingMetaForFirstItem(state: PagingState<Int, ConversationMessageIO>): PagingMetaIO? {
+    private suspend fun getPagingMetaForFirstItem(state: PagingState<String, ConversationMessageIO>): MatrixPagingMetaIO? {
         return state.pages.firstOrNull {
             it.data.isNotEmpty()
-        }?.data?.firstOrNull()?.conversationId?.let { id ->
+        }?.data?.firstOrNull()?.id?.let { id ->
             pagingMetaDao.getByEntityId(id)
         }
     }
 
     /** Returns paging meta data from the last item */
-    private suspend fun getPagingMetaForLastItem(state: PagingState<Int, ConversationMessageIO>): PagingMetaIO? {
+    private suspend fun getPagingMetaForLastItem(state: PagingState<String, ConversationMessageIO>): MatrixPagingMetaIO? {
         return state.pages.lastOrNull {
             it.data.isNotEmpty()
-        }?.data?.lastOrNull()?.conversationId?.let { id ->
+        }?.data?.lastOrNull()?.id?.let { id ->
             pagingMetaDao.getByEntityId(id)
         }
     }
